tools/qwen3vl-server/requirements.txt
torch
transformers
accelerate
qwen-vl-utils
flask
av
Pillow
